+++ hoomd/ParticleGroup.cc
@@ -97,9 +97,8 @@ ParticleGroup::ParticleGroup(std::shared_ptr<SystemDefinition> sysdef,
     sort(sorted_member_tags.begin(), sorted_member_tags.end());
 
     // store member tags
-    GlobalArray<unsigned int> member_tags_array(member_tags.size(), m_exec_conf);
+    GPUArray<unsigned int> member_tags_array(member_tags.size(), m_exec_conf);
     m_member_tags.swap(member_tags_array);
-    TAG_ALLOCATION(m_member_tags);
 
         {
         ArrayHandle<unsigned int> h_member_tags(m_member_tags,
@@ -110,20 +109,17 @@ ParticleGroup::ParticleGroup(std::shared_ptr<SystemDefinition> sysdef,
 
     // one byte per particle to indicate membership in the group, initialize with current number of
     // local particles
-    GlobalArray<unsigned int> is_member(m_pdata->getMaxN(), m_pdata->getExecConf());
+    GPUArray<unsigned int> is_member(m_pdata->getMaxN(), m_pdata->getExecConf());
     m_is_member.swap(is_member);
-    TAG_ALLOCATION(m_is_member);
 
-    GlobalArray<unsigned int> is_member_tag(m_pdata->getRTags().size(), m_pdata->getExecConf());
+    GPUArray<unsigned int> is_member_tag(m_pdata->getRTags().size(), m_pdata->getExecConf());
     m_is_member_tag.swap(is_member_tag);
-    TAG_ALLOCATION(m_is_member_tag);
 
     // build the reverse lookup table for tags
     buildTagHash();
 
-    GlobalArray<unsigned int> member_idx(member_tags.size(), m_pdata->getExecConf());
+    GPUArray<unsigned int> member_idx(member_tags.size(), m_pdata->getExecConf());
     m_member_idx.swap(member_idx);
-    TAG_ALLOCATION(m_member_idx);
 
     // now that the tag list is completely set up and all memory is allocated, rebuild the index
     // list
@@ -203,9 +199,8 @@ void ParticleGroup::updateMemberTags(bool force_update)
 #endif
 
         // store member tags in GlobalArray
-        GlobalArray<unsigned int> member_tags_array(member_tags.size(), m_pdata->getExecConf());
+        GPUArray<unsigned int> member_tags_array(member_tags.size(), m_pdata->getExecConf());
         m_member_tags.swap(member_tags_array);
-        TAG_ALLOCATION(m_member_tags);
 
         // sort member tags
         std::sort(member_tags.begin(), member_tags.end());
@@ -217,20 +212,17 @@ void ParticleGroup::updateMemberTags(bool force_update)
             std::copy(member_tags.begin(), member_tags.end(), h_member_tags.data);
             }
 
-        GlobalArray<unsigned int> member_idx(member_tags.size(), m_pdata->getExecConf());
+        GPUArray<unsigned int> member_idx(member_tags.size(), m_pdata->getExecConf());
         m_member_idx.swap(member_idx);
-        TAG_ALLOCATION(m_member_idx);
         }
 
     // one byte per particle to indicate membership in the group, initialize with current number of
     // local particles
-    GlobalArray<unsigned int> is_member(m_pdata->getMaxN(), m_pdata->getExecConf());
+    GPUArray<unsigned int> is_member(m_pdata->getMaxN(), m_pdata->getExecConf());
     m_is_member.swap(is_member);
-    TAG_ALLOCATION(m_is_member);
 
-    GlobalArray<unsigned int> is_member_tag(m_pdata->getRTags().size(), m_pdata->getExecConf());
+    GPUArray<unsigned int> is_member_tag(m_pdata->getRTags().size(), m_pdata->getExecConf());
     m_is_member_tag.swap(is_member_tag);
-    TAG_ALLOCATION(m_is_member_tag);
 
     // build the reverse lookup table for tags
     buildTagHash();
@@ -281,9 +273,8 @@ void ParticleGroup::reallocate()
     if (m_is_member_tag.getNumElements() != m_pdata->getRTags().size())
         {
         // reallocate if necessary
-        GlobalArray<unsigned int> is_member_tag(m_pdata->getRTags().size(), m_exec_conf);
+        GPUArray<unsigned int> is_member_tag(m_pdata->getRTags().size(), m_exec_conf);
         m_is_member_tag.swap(is_member_tag);
-        TAG_ALLOCATION(m_is_member_tag);
 
         buildTagHash();
         }