+++ hoomd/hpmc/IntegratorHPMCMono.h
@@ -174,7 +174,7 @@ template<class Shape> class IntegratorHPMCMono : public IntegratorHPMC
         }
 
     //! Get the interaction matrix
-    virtual const GlobalArray<unsigned int>& getInteractionMatrix()
+    virtual const GPUArray<unsigned int>& getInteractionMatrix()
         {
         return m_overlaps;
         }
@@ -338,10 +338,10 @@ template<class Shape> class IntegratorHPMCMono : public IntegratorHPMC
 
     protected:
     std::vector<param_type, hoomd::detail::managed_allocator<param_type>>
-        m_params;                         //!< Parameters for each particle type on GPU
-    GlobalArray<unsigned int> m_overlaps; //!< Interaction matrix (0/1) for overlap checks
-    detail::UpdateOrder m_update_order;   //!< Update order
-    bool m_image_list_is_initialized;     //!< true if image list has been used
+        m_params;                       //!< Parameters for each particle type on GPU
+    GPUArray<unsigned int> m_overlaps;  //!< Interaction matrix (0/1) for overlap checks
+    detail::UpdateOrder m_update_order; //!< Update order
+    bool m_image_list_is_initialized;   //!< true if image list has been used
     bool m_image_list_valid; //!< image list is invalid if the box dimensions or particle parameters
                              //!< have changed.
     std::vector<vec3<Scalar>>
@@ -405,9 +405,8 @@ IntegratorHPMCMono<Shape>::IntegratorHPMCMono(std::shared_ptr<SystemDefinition> 
         hoomd::detail::managed_allocator<param_type>(m_exec_conf->isCUDAEnabled()));
 
     m_overlap_idx = Index2D(m_pdata->getNTypes());
-    GlobalArray<unsigned int> overlaps(m_overlap_idx.getNumElements(), m_exec_conf);
+    GPUArray<unsigned int> overlaps(m_overlap_idx.getNumElements(), m_exec_conf);
     m_overlaps.swap(overlaps);
-    TAG_ALLOCATION(m_overlaps);
     ArrayHandle<unsigned int> h_overlaps(m_overlaps, access_location::host, access_mode::readwrite);
     for (unsigned int i = 0; i < m_overlap_idx.getNumElements(); i++)
         {
@@ -437,11 +436,6 @@ template<class Shape> void IntegratorHPMCMono<Shape>::update(uint64_t timestep)
     IntegratorHPMC::update(timestep);
 
     // get needed vars
-    ArrayHandle<hpmc_counters_t> h_counters(m_count_total,
-                                            access_location::host,
-                                            access_mode::readwrite);
-    hpmc_counters_t& counters = h_counters.data[0];
-
     const BoxDim box = m_pdata->getBox();
     unsigned int ndim = this->m_sysdef->getNDimensions();
 
@@ -487,6 +481,11 @@ template<class Shape> void IntegratorHPMCMono<Shape>::update(uint64_t timestep)
     // loop over local particles nselect times
     for (unsigned int i_nselect = 0; i_nselect < m_nselect; i_nselect++)
         {
+        ArrayHandle<hpmc_counters_t> h_counters(m_count_total,
+                                                access_location::host,
+                                                access_mode::readwrite);
+        hpmc_counters_t& counters = h_counters.data[0];
+
         // access particle data and system box
         ArrayHandle<Scalar4> h_postype(m_pdata->getPositions(),
                                        access_location::host,
@@ -788,8 +787,12 @@ template<class Shape> void IntegratorHPMCMono<Shape>::update(uint64_t timestep)
                 // Legacy external field energy difference
                 if (m_external)
                     {
-                    patch_field_energy_diff
-                        -= m_external->energydiff(timestep, i, pos_old, shape_old, pos_i, shape_i);
+                    patch_field_energy_diff -= m_external->energydiff(timestep,
+                                                                      h_tag.data[i],
+                                                                      pos_old,
+                                                                      shape_old,
+                                                                      pos_i,
+                                                                      shape_i);
                     }
 
                 // U_old - U_new