+++ hoomd/md/ComputeThermo.cc
@@ -30,9 +30,8 @@ ComputeThermo::ComputeThermo(std::shared_ptr<SystemDefinition> sysdef,
     m_exec_conf->msg->notice(5) << "Constructing ComputeThermo" << endl;
 
     assert(m_pdata);
-    GlobalArray<Scalar> properties(thermo_index::num_quantities, m_exec_conf);
+    GPUArray<Scalar> properties(thermo_index::num_quantities, m_exec_conf);
     m_properties.swap(properties);
-    TAG_ALLOCATION(m_properties);
 
     m_computed_flags.reset();
 
@@ -79,8 +78,8 @@ void ComputeThermo::computeProperties()
     ArrayHandle<unsigned int> h_tag(m_pdata->getTags(), access_location::host, access_mode::read);
 
     // access the net force, pe, and virial
-    const GlobalArray<Scalar4>& net_force = m_pdata->getNetForce();
-    const GlobalArray<Scalar>& net_virial = m_pdata->getNetVirial();
+    const GPUArray<Scalar4>& net_force = m_pdata->getNetForce();
+    const GPUArray<Scalar>& net_virial = m_pdata->getNetVirial();
     ArrayHandle<Scalar4> h_net_force(net_force, access_location::host, access_mode::read);
     ArrayHandle<Scalar> h_net_virial(net_virial, access_location::host, access_mode::read);